look #1

+++ src/main/java/gregtech/GT_Mod.java
@@ -330,7 +330,7 @@ public class GT_Mod
                 List<ModContainer> tModList = tLoadController.getActiveModList();
                 List<ModContainer> tNewModsList = new ArrayList();
                 ModContainer tGregTech = null;
-                short tModList_sS=tModList.size();
+                short tModList_sS=(short) tModList.size();
                 for (short i = 0; i < tModList_sS; i = (short) (i + 1)) {
                     ModContainer tMod = (ModContainer) tModList.get(i);
                     if (tMod.getModId().equalsIgnoreCase("gregtech")) {
